+++ drawing.c
@@ -1,0 +1,56 @@
+#include <mlx.h>
+#include "stone.h"
+
+void my_mlx_pixel_put(t_data *data, int x, int y, int color)
+{
+    char *dst;
+    dst = data->addr + (y * data->line_length + x * (data->bits_per_pixel / 8));
+    *(unsigned int *)dst = color;
+}
+
+void draw_line(t_data *data, int x0, int y0, int x1, int y1, int color)
+{
+    int dx, dy;
+    int sx, sy;
+    int err;
+
+    if (x0 < x1)
+    {
+        dx = x1 - x0;
+        sx = 1;
+    }
+    else
+    {
+        dx = x0 - x1;
+        sx = -1;
+    }
+    if (y0 < y1)
+    {
+        dy = y1 - y0;
+        sy = 1;
+    }
+    else
+    {
+        dy = y0 - y1;
+        sy = -1;
+    }
+    err = dx - dy;
+
+    while (x0 != x1 && y0 != y1)
+    {
+        int e2;
+
+        my_mlx_pixel_put(data, x0, y0, color);
+        e2 = 2 * err;
+        if (e2 > -dy)
+        {
+            err -= dy;
+            x0 += sx;
+        }
+        if (e2 < dx)
+        {
+            err += dx;
+            y0 += sy;
+        }
+    }
+}
